requirements.txt
torch
torchvision
diffusers
transformers
accelerate
mmdet
mmengine
opencv-python
numpy
pyyaml
tqdm
scipy
timm